+++ routers/userRouter.mjs
@@ -12,7 +12,7 @@ const userCtrl = new UserCtrl('main', db.User, db);
 // router.post('/logout', userCtrl.getMain.bind(userCtrl));
 router.post('/register', userCtrl.postRegister.bind(userCtrl));
 router.post('/login', userCtrl.postLogin.bind(userCtrl));
-router.post('/email', userCtrl.postEmail.bind(userCtrl));\
+router.post('/email', userCtrl.postEmail.bind(userCtrl));
 
 // routes for friends page functionality
 router.post('/friends', userCtrl.addFriends.bind(userCtrl));
